requirements_dev.txt
coverage~=5.5
django-stubs~=1.8.0
Faker~=9.7.1
lorem~=0.1.1
mixer~=7.2.0
mock~=4.0.3
mypy~=0.901
pbr~=5.6.0
pre-commit~=2.15.0
pylint~=2.11.1
pylint-django~=2.4.4
pytest~=6.2.5
pytest-cov~=3.0.0
pytest-django~=4.4.0
types-python-dateutil~=0.1.2
